move ui code from __init__
